requirements.txt
dj-rest-auth==4.0.1
Django==4.2.3
django-allauth==0.54.0
django-cors-headers==4.2.0
django-extensions==3.2.3
django-filter==23.2
django-model-utils==4.3.1
django-rest-auth==0.9.5
django-s3-storage==0.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
fcm-django==2.0.0
Pillow==10.0.0
PyJWT==2.8.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
six==1.16.0
